+++ scripts/retrieve_additional_tyndp_data.py
@@ -2,16 +2,24 @@
 #
 # SPDX-License-Identifier: MIT
 """
-The TYNDP PECD data contains input data for the 2024 TYNDP scenario building process.
-
-This rule downloads the TYNDP PECD v3.1 data from Google Drive and extracts it in the ``data/tyndp_2024_bundle``
+Retrieves additional TYNDP data not included in the Zenodo TYNDP data bundle.
+Downloads the zip file from Google Drive and extracts it in the ``data/tyndp_2024_bundle``
 subdirectory, such that all files of the TYNDP bundle are stored in it.
 The original data is published by ENTSO-E and ENTSOG under Creative Commons Attribution 4.0 International License (CC-BY 4.0)
 and can be found under https://2024.entsos-tyndp-scenarios.eu/download/.
 
+Currently, this is used for two additional datasets:
+* TYNDP PECD data: The TYNDP PECD v3.1 data contains input data for the 2024 TYNDP scenario building process.
+* TYNDP hydro inflows: The TYNDP hydro inflow data from PEMMDB v2.4 contains hydro inflow data for different hydro technologies:
+    * Run of River,
+    * Pondage,
+    * Reservoir,
+    * PS Open,
+    * PS Closed
+
 **Outputs**
 
-- ``data/tyndp_2024_bundle/PECD``: PECD input data for TYNDP 2024 scenario building
+- ``data/tyndp_2024_bundle/<additional-dataset>``: Additional input dataset for TYNDP 2024 scenario building
 
 """
 
@@ -24,9 +32,7 @@ from _helpers import configure_logging, progress_retrieve, set_scenario_config
 
 logger = logging.getLogger(__name__)
 
-# Define the base URL
-# TODO: retrieve_tyndp_pecd_data needs to be deprecated once PECD data is added to the TYNDP data bundle
-url = "https://storage.googleapis.com/open-tyndp-data-store/PECD.zip"
+# TODO: retrieve_additional_tyndp_data needs to be deprecated once all TYNDP data is added to the TYNDP data bundle
 
 if __name__ == "__main__":
     if "snakemake" not in globals():
@@ -41,20 +47,24 @@ if __name__ == "__main__":
     set_scenario_config(snakemake)
     disable_progress = snakemake.config["run"].get("disable_progressbar", False)
 
+    source = snakemake.params.source
     to_fn = snakemake.output.dir
     tyndp_bundle_fn = Path(rootpath, snakemake.params["tyndp_bundle"])
     to_fn_zp = to_fn + ".zip"
 
+    # define url
+    url = snakemake.params.url
+
     # download .zip file
-    logger.info(f"Downloading TYNDP PECD data from '{url}'.")
+    logger.info(f"Downloading TYNDP {source} data from '{url}'.")
     progress_retrieve(url, to_fn_zp, disable=disable_progress)
 
     # extract
-    logger.info("Extracting TYNDP PECD data.")
+    logger.info(f"Extracting TYNDP {source} data.")
     with zipfile.ZipFile(to_fn_zp, "r") as zip_ref:
         zip_ref.extractall(tyndp_bundle_fn)
 
     # remove .zip file
     os.remove(to_fn_zp)
 
-    logger.info(f"TYNDP PECD data available in '{to_fn}'.")
+    logger.info(f"TYNDP {source} data available in '{to_fn}'.")